// src/aliexpress/dto/oauth.dto.ts
export interface AliAccessTokenDTO {
  refresh_token_valid_time: number;
  havana_id: string;
  expire_time: number;
  locale: string;
  user_nick: string;
  access_token: string;
  refresh_token: string;
  user_id: string;
  account_platform: string;
  refresh_expires_in: number;
  expires_in: number;
  sp: string;
  seller_id: string;
  account: string;
  code: string;
  request_id: string;
  _trace_id_: string;
  message: string;
}